+++ sys8051/net/gnrc/network_layer/icmpv6/echo/gnrc_icmpv6_echo.c
@@ -29,7 +29,7 @@
 #endif
 /* 8051 implementation */
 gnrc_pktsnip_t *gnrc_icmpv6_echo_build(uint8_t type, uint16_t id, uint16_t seq,
-                                       uint8_t *data, size_t data_len)
+                                       uint8_t *data, uint32_t data_len)
 {
     gnrc_pktsnip_t *pkt;
     icmpv6_echo_t *echo;
